requirements.txt
googletrans
streamlit
requests-html
selenium